Move schema constraints into TypeScript annotations for new schema generation
