+++ src/schema/bin.schema.ts
@@ -1,23 +1,42 @@
 import {QUANTITATIVE} from '../type';
 import {toMap} from '../util';
 
-/** binning properties or boolean flag for determining whether to bin data or not. */
+/**
+ * Binning properties or boolean flag for determining whether to bin data or not.
+ */
 export interface BinProperties {
-  /** The minimum bin value to consider. If unspecified, the minimum value of the specified field is used. */
+  /**
+   * The minimum bin value to consider. If unspecified, the minimum value of the specified field is used.
+   */
   min?: number;
-  /** The maximum bin value to consider. If unspecified, the maximum value of the specified field is used. */
+  /**
+   * The maximum bin value to consider. If unspecified, the maximum value of the specified field is used.
+   */
   max?: number;
-  /** The number base to use for automatic bin determination (default is base 10). */
+  /**
+   * The number base to use for automatic bin determination (default is base 10).
+   */
   base?: number;
-  /** An exact step size to use between bins. If provided, options such as maxbins will be ignored. */
+  /**
+   * An exact step size to use between bins. If provided, options such as maxbins will be ignored.
+   */
   step?: number;
-  /** An array of allowable step sizes to choose from. */
+  /**
+   * An array of allowable step sizes to choose from.
+   */
   steps?: number[];
-  /** A minimum allowable step size (particularly useful for integer values). */
+  /**
+   * A minimum allowable step size (particularly useful for integer values).
+   */
   minstep?: number;
-  /** Scale factors indicating allowable subdivisions. The default value is [5, 2], which indicates that for base 10 numbers (the default base), the method may consider dividing bin sizes by 5 and/or 2. For example, for an initial step size of 10, the method can check if bin sizes of 2 (= 10/5), 5 (= 10/2), or 1 (= 10/(5*2)) might also satisfy the given constraints. */
+  /**
+   * Scale factors indicating allowable subdivisions. The default value is [5, 2], which indicates that for base 10 numbers (the default base), the method may consider dividing bin sizes by 5 and/or 2. For example, for an initial step size of 10, the method can check if bin sizes of 2 (= 10/5), 5 (= 10/2), or 1 (= 10/(5*2)) might also satisfy the given constraints.
+   */
   div?: number[];
-  /** Maximum number of bins. */
+  /**
+   * Maximum number of bins.
+   * @minimum 2
+   */
   maxbins?: number;
 }
 
@@ -49,8 +68,7 @@ export var binProperties = {
         type: 'array'
       },
       maxbins: {
-        type: 'integer',
-        minimum: 2
+        type: 'integer'
       }
     },
   }, {
