+++ src/schema/data.schema.ts
@@ -1,7 +1,13 @@
 export interface Data {
+
+  /**
+   * @enum ["json", "csv", "tsv"]
+   */
   formatType?: string;
   url?: string;
-  /** Pass array of objects instead of a url to a file. */
+  /**
+   * Pass array of objects instead of a url to a file.
+   */
   values?: any[];
 }
 
@@ -10,8 +16,7 @@ export var data = {
   properties: {
     // data source
     formatType: {
-      type: 'string',
-      enum: ['json', 'csv', 'tsv']
+      type: 'string'
     },
     url: {
       type: 'string'
