+++ src/schema/schema.ts
@@ -8,9 +8,13 @@ import {encoding, Encoding} from './encoding.schema';
 import {Mark} from '../mark';
 import {transform, Transform} from './transform.schema';
 
-/** Schema for Vega-Lite specification */
+/**
+ * Schema for Vega-Lite specification
+ */
 export interface Spec {
-  /** A name for the specification. The name is used to annotate marks, scale names, and more. */
+  /**
+   * A name for the specification. The name is used to annotate marks, scale names, and more.
+   */
   name?: string;
   description?: string;
   data?: Data;
@@ -26,7 +30,7 @@ export {aggregate} from './fielddef.schema';
 export var util = schemaUtil;
 
 /** @type Object Schema of a vega-lite specification */
-export var schema = {
+export var schemaOLD = {
   $schema: 'http://json-schema.org/draft-04/schema#',
   type: 'object',
   required: ['mark', 'encoding'],
@@ -43,9 +47,4 @@ export var schema = {
     encoding: encoding,
     config: config
   }
-};
-
-/** Instantiate a verbose vl spec from the schema */
-export function instantiate() {
-  return schemaUtil.instantiate(schema);
 };
