--- src/schema/schemagen.ts
@@ -1,3 +1,0 @@
-import {schema} from './schema';
-
-process.stdout.write(JSON.stringify(schema, null, 4) + '\n');
