+++ src/schema/sort.schema.ts
@@ -5,10 +5,23 @@ import {toMap} from '../util';
 export type SortEnum = string; // TODO: string literal "ascending", "descending", "none"
 
 export interface SortField {
-  /** The field name to aggregate over. */
+  /**
+   * The field name to aggregate over.
+   */
   field: string;
-  /** The field name to aggregate over. */
+  /**
+   * The field name to aggregate over.
+   * @enum ["values", "count", "valid", "missing", "distinct",
+   *   "sum", "mean", "average", "variance", "variancep", "stdev",
+   *   "stdevp", "median", "q1", "q3", "modeskew", "min", "max",
+   *   "argmin", "argmax"]
+   *
+   */
   op: string;
+
+  /**
+   * @enum: ["ascending", "descending"]
+   */
   order?: string;
 }
 
@@ -25,12 +38,10 @@ const sortField = { // sort by aggregation of another field
       type: 'string'
     },
     op: {
-      type: 'string',
-      enum: AGGREGATE_OPS
+      type: 'string'
     },
     order: {
-      type: 'string',
-      enum: ['ascending', 'descending']
+      type: 'string'
     }
   }
 };
